Classes were separated into separate files
