+++ modules/Calendar/iCal/IcalendarJournal.php
@@ -1,0 +1,8 @@
+<?php
+// $Id: IcalendarJournal.php,v 1.8 2005/07/21 22:31:44 defacer Exp $
+require_once('include/utils/utils.php');
+
+class IcalendarJournal extends IcalendarComponent
+{
+
+}
